refactor: move data processing to correct location
