+++ src/psycop/common/model_evaluation/binary/time/timedelta_data.py
@@ -3,6 +3,7 @@ from typing import Literal
 
 import numpy as np
 import pandas as pd
+import polars as pl
 from psycop.common.model_evaluation.binary.utils import (
     auroc_by_group,
     sensitivity_by_group,
@@ -220,3 +221,43 @@ def get_sensitivity_by_timedelta_df(
         groupby_col_name="unit_from_event_binned",
         confidence_interval=confidence_interval,
     )
+
+
+def get_time_from_first_positive_to_diagnosis_df(
+    input_df: pd.DataFrame,
+) -> pd.DataFrame:
+    """input_df must contain columns:
+        y: 1 or 0
+        pred: 1 or 0
+        id: Patient ID
+        pred_timestamps: Timestamp of prediction
+        outcome_timestamps: Timestamp of outcome
+    Returns a dataframe with the additional columns:
+        years_from_pred_to_event: Years from first positive prediction to event
+        days_from_pred_to_event: Days from first positive prediction to event
+    """
+    df = pl.from_pandas(input_df).with_columns(
+        (pl.col("outcome_timestamps") - pl.col("pred_timestamps")).alias(
+            "time_from_pred_to_event"
+        )
+    )
+
+    ever_positives = df.filter(
+        pl.col("time_from_pred_to_event").is_not_null() & pl.col("pred") == 1
+    )
+
+    plot_df = (
+        ever_positives.sort("time_from_pred_to_event", descending=True)
+        .groupby("id")
+        .head(1)
+        .with_columns(
+            (pl.col("time_from_pred_to_event").dt.days() / 365.25).alias(
+                "years_from_pred_to_event",
+            ),
+            (pl.col("time_from_pred_to_event").dt.days()).alias(
+                "days_from_pred_to_event",
+            ),
+        )
+    ).to_pandas()
+
+    return plot_df
