+++ src/psycop/projects/t2d/paper_outputs/model_description/performance/incidence_by_time_until_diagnosis.py
@@ -1,6 +1,8 @@
 import pandas as pd
 import plotnine as pn
-import polars as pl
+from psycop.common.model_evaluation.binary.time.timedelta_data import (
+    get_time_from_first_positive_to_diagnosis_df,
+)
 from psycop.projects.t2d.paper_outputs.config import (
     PN_THEME,
 )
@@ -17,7 +19,7 @@ def t2d_first_pred_to_event(run: PipelineRun) -> pn.ggplot:
             )[0],
             "y": eval_ds.y,
             "id": eval_ds.ids,
-            "pred_timestamp": eval_ds.pred_timestamps,
+            "pred_timestamps": eval_ds.pred_timestamps,
             "outcome_timestamps": eval_ds.outcome_timestamps,
         },
     )
@@ -49,47 +51,6 @@ def t2d_first_pred_to_event(run: PipelineRun) -> pn.ggplot:
     p.save(run.paper_outputs.paths.figures / "first_pred_to_event.png")
 
     return p
-
-
-def get_time_from_first_positive_to_diagnosis_df(
-    input_df: pd.DataFrame,
-) -> pd.DataFrame:
-    """input_df must contain columns:
-        y: 1 or 0
-        pred: 1 or 0
-        id: Patient ID
-        pred_timestamps: Timestamp of prediction
-        outcome_timestamps: Timestamp of outcome
-
-    Returns a dataframe with the additional columns:
-        years_from_pred_to_event: Years from first positive prediction to event
-        days_from_pred_to_event: Days from first positive prediction to event
-    """
-    df = pl.from_pandas(input_df).with_columns(
-        (pl.col("outcome_timestamps") - pl.col("pred_timestamps")).alias(
-            "time_from_pred_to_event"
-        )
-    )
-
-    ever_positives = df.filter(
-        pl.col("time_from_pred_to_event").is_not_null() & pl.col("pred") == 1
-    )
-
-    plot_df = (
-        ever_positives.sort("time_from_pred_to_event", descending=True)
-        .groupby("id")
-        .head(1)
-        .with_columns(
-            (pl.col("time_from_pred_to_event").dt.days() / 365.25).alias(
-                "years_from_pred_to_event",
-            ),
-            (pl.col("time_from_pred_to_event").dt.days()).alias(
-                "days_from_pred_to_event",
-            ),
-        )
-    ).to_pandas()
-
-    return plot_df
 
 
 if __name__ == "__main__":
